pyproject.toml
[project]
dependencies = [
  "cryptography>=46.0.0",
  "fastmcp>=2.14",
  "pydantic>=2.11.7",
  "pydantic-settings>=2.10.1",
]
description = "Add your description here"
name = "lampyrid"
readme = "README.md"
requires-python = ">=3.14"
version = "0.2.0"

[project.scripts]
lampyrid = "lampyrid.__main__:main"

[dependency-groups]
dev = [
  "datamodel-code-generator>=0.52.0",
  "dirty-equals>=0.11",
  "inline-snapshot>=0.31.1",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "python-dotenv>=1.2.1",
  "ruff>=0.14.11",
  "ty>=0.0.12",
]

[build-system]
build-backend = "uv_build"
requires = ["uv_build"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
# "I" (isort), "D" (pydocstyle), "E" (pycodestyle)
extend-select = ["I", "D", "E"]
ignore = [
  "D203", # 1 blank line required before class docstring (conflicts with D211)
  "D213", # Multi-line docstring summary should start at the second line (conflicts with D212)
]

[tool.ruff.lint.per-file-ignores]
"src/lampyrid/models/firefly_models.py" = ["D100", "D101", "E501"]
"tests/**/*.py" = ["D104"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
  "integration: marks tests as integration tests",
  "accounts: account-related tests",
  "transactions: transaction-related tests",
  "budgets: budget-related tests",
]

[tool.datamodel-codegen]
input = "firefly-iii-6.4.14-v1.yaml"
output = "src/lampyrid/models/firefly_models.py"
output-model-type = "pydantic_v2.BaseModel"
